splits patch logic out of flowerful
