+++ flowerful.ts
@@ -2,17 +2,13 @@
  * Flowerful runtime detour library for Creator of Another World
  */
 
-//#region typeDefs
-
 import type { FlowerAPI } from "./api/FlowerAPI";
-import type { FlowerPatch, Patchable, PatchFn } from "./api/FlowerPatch"
 import type { FlowerPlugin } from "./api/FlowerPlugin"
-import { LogSource } from "./logSource"
+import { LogSource } from "./api/logSource";
+import { ApplyAllPatches, RegisterPatch } from "./flowerful.patches";
 
 /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
 declare const nw: any;
-
-//#endregion typeDefs
 
 //#region flower_ctor
 
@@ -76,11 +72,7 @@ async function LoadAllPlugins()
         Plugins[guid].PluginAwake();
     }
 
-    //Apply patches
-    for (const patch of patches)
-    {
-        Apply(patch)
-    }
+    ApplyAllPatches();
 }
 
 async function LoadPlugin(file: string)
@@ -112,7 +104,7 @@ async function LoadPlugin(file: string)
             Plugins[plugin.GUID] = plugin;
 
             //Tell the plugin it is being registered and pass it the API/Logger
-            plugin.PluginRegistered(flowerAPI, new LogSource(plugin.GUID));
+            plugin.PluginRegistered(flowerAPI, new LogSource(plugin.GUID, WriteLog));
         }
         else
         {
@@ -170,97 +162,6 @@ function onLoggerWindowLoaded(win: any)
 }
 
 //#endregion flower-logger
-
-//#region flower-patcher
-
-const patches: FlowerPatch[] = [];
-
-function FindPatch(obj: Patchable, method: string)
-{
-    if (!obj[method])
-    {
-        console.error(`Method ${method} not found on ${obj}`);
-        return;
-    }
-
-    for (const patch of patches)
-    {
-        if (patch.obj === obj && patch.methodName === method)
-        {
-            return patch;
-        }
-    }
-
-    const patch: FlowerPatch = {
-        obj: obj,
-        methodName: method,
-        prefixes: [],
-        postfixes: []
-    }
-
-    patches.push(patch);
-    return patch;
-}
-
-function Apply(patch: FlowerPatch)
-{
-    /* eslint-disable-next-line @typescript-eslint/ban-types */
-    const orig = patch.obj[patch.methodName] as Function;
-
-    const wrapper: PatchFn = function (...args)
-    {
-        WriteLog("Flower", `Running detour for ${patch.methodName}`);
-        // <-- this = obj
-
-        WriteLog("Flower", `Prefixes ${patch.prefixes.length}`);
-        //patch.prefixes.forEach(prefix => prefix.call(patch.obj, ...args));
-        //Allow ending the detour early
-        for (const prefix of patch.prefixes)
-        {
-            if (false === prefix.call(patch.obj, ...args))
-            {
-                WriteLog("Flower", "Ending detour");
-                return;
-            }
-        }
-
-        try
-        {
-            orig.call(patch.obj, ...args);
-        }
-        catch (e)
-        {
-            WriteLog("Flower", `Error running orig: ${e}`)
-            return;
-        }
-
-        WriteLog("Flower", `Postfixes ${patch.postfixes.length}`);
-        patch.postfixes.forEach(postfix => postfix.call(patch.obj, ...args));
-    }
-
-    patch.obj[patch.methodName] = wrapper.bind(patch.obj);
-}
-
-function RegisterPatch(obj: Patchable, methodName: string, patch: PatchFn, isPrefix: boolean)
-{
-
-    WriteLog("Flower", `Running RegisterPatch for ${methodName}`);
-
-    const accum = FindPatch(obj, methodName);
-    if (!accum) return false;
-
-    if (isPrefix)
-    {
-        accum.prefixes.push(patch);
-    }
-    else
-    {
-        accum.postfixes.push(patch);
-    }
-
-    return true;
-}
-//#endregion flower-patcher
 
 //Internal Context
 document._flowerInt = { Init }
